% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.dat.R
\docType{data}
\name{demo.dat}
\alias{demo.dat}
\title{Demo Data}
\format{
A data frame with 3000 rows and 7 columns.
\describe{
\item{X1}{Observed variable 1.}
\item{X2}{Observed variable 2.}
\item{X3}{Observed variable 3.}
\item{X4}{Observed variable 4.}
\item{X5}{Observed variable 5.}
\item{id}{ID variable for each individual (1 to 8).}
\item{time}{Time variable (1 to 300).}
}
}
\usage{
demo.dat
}
\description{
Demo Data
}
\keyword{data}
