% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.dynr.var.R
\docType{data}
\name{demo.dynr.var}
\alias{demo.dynr.var}
\title{Demo Output of dynr.var()}
\format{
A list of length ten. Each element of the list has the followin elemens
\describe{
\item{Betas}{Maybe we can put a description here.}
\item{ResidCov}{Maybe we can put a description here.}
\item{PCC}{Maybe we can put a description here.}
\item{PDC}{Maybe we can put a description here.}
\item{Res}{Maybe we can put a description here.}
}
}
\usage{
demo.dynr.var
}
\description{
Demo Output of dynr.var()
}
\keyword{data}
