% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynr.cookdist.R
\name{dynr.cookdist}
\alias{dynr.cookdist}
\title{Cook's Distance in Dynr}
\usage{
dynr.cookdist(dataframe, id, time, ini.beta, dir = getwd())
}
\arguments{
\item{dataframe}{a data frame object of data that contain a column of subject ID numbers (i.e., an ID variable), a column indicating subject-specific measurement occasions (i.e., a TIME variable), at least one column of observed values, and any number of covariates. The TIME variable should contain subject-specific sequences of (subsets of) consecutively equally spaced numbers (e.g, 1, 2, 3, ...). That is, the program assumes that the input data.frame is equally spaced with potential missingness. If the measurement occasions for a subject are a subset of an arithmetic sequence but are not consecutive, NAs will be inserted automatically to create an equally spaced data set before estimation. Missing values in the observed variables shoud be indicated by NA.}

\item{id}{a character string of the name of the ID variable in the data.}

\item{time}{a character string of the name of the TIME variable in the data.}

\item{ini.beta}{the matrix of starting/fixed values for the transition matrix in the specified linear dynamic model.}

\item{dir}{path for output files.}
}
\description{
Cook's Distance in Dynr
}
\references{
Add references to the Cook's distance applied to dynr here.
}
\author{
Jonathan Park

Maybe we can put a description here on how Cook's distance is calculated.
Also Jonathan can you give me the data set you used in the example?

I made the following changes to the code.
\enumerate{
\item I changed the name to \code{dynr.cookdist} to match the other functions in the package.
\item I changed the argument \code{data} to \code{dataframe} to match \code{\link[dynr:dynr.data]{dynr::dynr.data()}}.
\item I changed the argument \code{ID} to \code{id} to match \code{\link[dynr:dynr.data]{dynr::dynr.data()}}.
\item I set a default value for the argument \code{dir}.
\item I used \code{\link[styler:style_pkg]{styler::style_pkg()}} to style the code for added readability.
}
}
