% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynr.sub.R
\name{dynr.sub}
\alias{dynr.sub}
\title{Subgrouping Algorithm}
\usage{
dynr.sub(
  input,
  alpha = 0.05,
  params.cook = NULL,
  params.var = c("Betas", "ResidCov", "PDC", "PCC", "Both"),
  method = c("hard", "fuzz"),
  k = 2,
  m = 2
)
}
\arguments{
\item{input}{list of \code{\link[dynr:dynr.cook]{dynr::dynr.cook()}} objects or output from \code{\link[=dynr.var]{dynr.var()}}.}

\item{alpha}{significance level for testing of transition matrix coefficients.}

\item{params.cook}{parameters for clustering for \code{\link[dynr:dynr.cook]{dynr::dynr.cook()}} input.}

\item{params.var}{parameters for clustering for \code{\link[=dynr.var]{dynr.var()}} input.}

\item{method}{Subgrouping method.}

\item{k}{number of clusters.}

\item{m}{value of fuzzifier.}
}
\description{
Subgrouping Algorithm
}
\details{
I encapsulated \code{dynrautoVAR} into a temporary package so I can render the documentation and do some tests.

I made the following changes to the code.
\enumerate{
\item I changed the name to \code{dynr.sub()} to match the other functions in the package.
\item I changed the argument \code{inputs} to \code{input}.
\item I removed the argument \code{type} and based the type on the class of the input.
\item I changed the argument \code{p.val} to \code{alpha}.
\item I changed the argument \code{params} to \code{params.cook} and \verb{var.opt`` to }params.var`.
\item I used \code{\link[styler:style_pkg]{styler::style_pkg()}} to style the code for added readability.
}
}
\author{
Jonathan Park

This function takes a list of outputs from \code{\link[dynr:dynr.cook]{dynr::dynr.cook()}} or output from \code{\link[=dynr.var]{dynr.var()}} and
performs a subgrouping analysis. The output is a list of clusters.
}
