% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynr.var.R
\name{dynr.var}
\alias{dynr.var}
\title{Automatically Fit VAR(1) Models Based on Input Data}
\usage{
dynr.var(
  dataframe,
  nv,
  id = NULL,
  time,
  dir = getwd(),
  alpha = 0.05,
  ini.mu = NULL,
  ini.cov = NULL,
  ini.beta = NULL,
  ini.loadings = NULL,
  use.mi = FALSE,
  aux = NULL
)
}
\arguments{
\item{dataframe}{a data frame object of data that contain a column of subject ID numbers (i.e., an ID variable), a column indicating subject-specific measurement occasions (i.e., a TIME variable), at least one column of observed values, and any number of covariates. The TIME variable should contain subject-specific sequences of (subsets of) consecutively equally spaced numbers (e.g, 1, 2, 3, ...). That is, the program assumes that the input data.frame is equally spaced with potential missingness. If the measurement occasions for a subject are a subset of an arithmetic sequence but are not consecutive, NAs will be inserted automatically to create an equally spaced data set before estimation. Missing values in the observed variables shoud be indicated by NA.}

\item{nv}{number of variables.}

\item{id}{a character string of the name of the ID variable in the data.}

\item{time}{a character string of the name of the TIME variable in the data.}

\item{dir}{path for output files.}

\item{alpha}{significance level for testing of transition matrix coefficients.}

\item{ini.mu}{a vector of the starting or fixed values of the initial state vector.}

\item{ini.cov}{a positive definite matrix of the starting or fixed values of the initial error covariance structure.}

\item{ini.beta}{the matrix of starting/fixed values for the transition matrix in the specified linear dynamic model.}

\item{ini.loadings}{matrix of starting or fixed values for factor loadings.}

\item{use.mi}{if \code{use.mi = TRUE}, use \code{\link[dynr:dynr.mi]{dynr::dynr.mi()}} to address missing data in covariates.}

\item{aux}{names of the auxiliary variables used in the imputation model
if \code{use.mi = TRUE}.}
}
\value{
Returns a list of fitted VAR(1) models for each subject.
}
\description{
Automatically Fit VAR(1) Models Based on Input Data
}
\details{
I encapsulated \code{dynrautoVAR} into a temporary package so I can render the documentation and do some tests.

I made the following changes to the code.
\enumerate{
\item I changed the argument \code{data} to \code{dataframe} to match \code{\link[dynr:dynr.data]{dynr::dynr.data()}}.
\item I changed the argument \code{ID} to \code{id} to match \code{\link[dynr:dynr.data]{dynr::dynr.data()}}.
\item I changed the argument \code{p.val} to \code{alpha}.
\item I removed the default \code{NULL} value on arguments that require explicit values (\code{dataframe}, \code{nv}, and \code{time}).
\item I made all initial condition arguments begin with \code{ini.}.
\item I set a default value for the argument \code{dir} to remove it from the argument handling section.
\item I cleaned up the argument handling section.
\item I changed \code{result$estimation.result} to \code{results$estimation.result}.
\item I gave the output a class of \code{dynrVar} for subsequent use of the output.
\item I used \code{\link[styler:style_pkg]{styler::style_pkg()}} to style the code for added readability.
}

Note that I mainly based the way I documented the arguments from functions in the \code{dynr} package used within this function. If you are amenable to the way I documented this function, I will proceed to documenting the rest.
}
\references{
Add references to \code{\link[=dynr.var]{dynr.var()}} here.
}
\author{
Jonathan Park

This function takes in a \code{data.frame}, fits individual models for each subject indicated by the ID variable
and stores the fitted models as a list.
Note that if \verb{ini.*} arguments are not provided, initial values are set to null matrices or null vectors.
}
