% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynrautoVAR-package.R
\docType{package}
\name{dynrautoVAR-package}
\alias{dynrautoVAR-package}
\alias{_PACKAGE}
\title{dynrautoVAR: Automated VAR(1) in dynr}
\description{
Automatically fit VAR(1) models based on input data in 'dynr'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jeksterslab/dynrautoVAR}
  \item \url{https://jeksterslab.github.io/dynrautoVAR/}
  \item Report bugs at \url{https://github.com/jeksterslab/dynrautoVAR/issues}
}

}
\author{
\strong{Maintainer}: Jonathan Park \email{JPark@psu.edu} [copyright holder]

}
\keyword{internal}
